% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cors.to.q}
\alias{cors.to.q}
\title{Conversion from a correlation Difference to Cohen's q}
\usage{
cors.to.q(rho1, rho2, verbose = 1)
}
\arguments{
\item{rho1}{first correlation.}

\item{rho2}{second correlation.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and definitions), if \code{0} no output is printed
on the console.}
}
\value{
\item{q}{Cohen's q effect size.}
\item{delta}{correlation difference: rho1 - rho2.}
\item{rho1}{first correlation.}
\item{rho2}{second correlation.}
}
\description{
Conversion from a correlation Difference to Cohen's q
}
\examples{
cors.to.q(rho2 = 0.5712027, rho1 = 0.50)
cors.to.q(rho2 = 0.6907068, rho1 = 0.50)
cors.to.q(rho2 = 0.7815365, rho1 = 0.50)

}
