% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{etasq.to.f}
\alias{etasq.to.f}
\title{Conversion from Eta-squared to Cohen's f}
\usage{
etasq.to.f(eta.squared, verbose = 1)
}
\arguments{
\item{eta.squared}{(Partial) Eta-squared.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}
}
\value{
\item{f}{Cohen's f.}
\item{f.squared}{Cohen's f².}
\item{eta.squared}{(Partial) Eta-squared.}
}
\description{
Conversion from Eta-squared to Cohen's f
}
\examples{
etasq.to.f(eta.squared = 0.01) # small
etasq.to.f(eta.squared = 0.06) # medium
etasq.to.f(eta.squared = 0.14) # large

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences}
(2nd ed.). Lawrence Erlbaum Associates.
}
