% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f.to.etasq}
\alias{f.to.etasq}
\title{Conversion between Cohen's f and Eta-squared}
\usage{
f.to.etasq(f, verbose = 1)
}
\arguments{
\item{f}{Cohen's f.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}
}
\value{
\item{eta.squared}{(Partial) Eta-squared.}
\item{f.squared}{Cohen's f².}
\item{f}{Cohen's f.}
}
\description{
Conversion between Cohen's f and Eta-squared
}
\examples{
f.to.etasq(f = 0.10) # small
f.to.etasq(f = 0.25) # medium
f.to.etasq(f = 0.40) # large

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences}
(2nd ed.). Lawrence Erlbaum Associates.
}
