% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inflate.sample}
\alias{inflate.sample}
\title{Inflate Sample Size for Attrition}
\usage{
inflate.sample(n, rate = 0.05, ceil.n = TRUE, verbose = 1)
}
\arguments{
\item{n}{sample size.}

\item{rate}{attrition rate.}

\item{ceil.n}{rounds-up the inflated sample size.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and definitions), if \code{0} no output is printed
on the console.}
}
\value{
inflated sample size.
}
\description{
Inflate Sample Size for Attrition
}
\examples{
inflate.sample(n = 100, rate = 0.05)

}
