% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{means.to.d}
\alias{means.to.d}
\title{Conversion from Means and Standard Deviations to Cohen's d}
\usage{
means.to.d(
  mu1,
  mu2 = 0,
  sd1 = 1,
  sd2 = 1,
  n.ratio = 1,
  n2 = 10000000000,
  paired = FALSE,
  rho.paired = 0.5,
  verbose = 1
)
}
\arguments{
\item{mu1}{mean of the first group.}

\item{mu2}{mean of the second group.}

\item{sd1}{standard deviation of the first group.}

\item{sd2}{standard deviation of the second group.}

\item{n.ratio}{\code{n1 / n2} ratio (applies to independent samples only).}

\item{n2}{integer; sample size in the second group (or for the single group
in paired samples).}

\item{paired}{if \code{TRUE} paired samples}

\item{rho.paired}{correlation between repeated measures for paired samples
(e.g., pretest and post-test).}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key parameters and
definitions), if \code{0} no output is printed on the console.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{d}{Cohen's d}
\item{pooled.sd}{Pooled standard deviation}
\item{var.ratio}{Ratio of the variance in the two groups (applies to independent samples only)}
\item{n1}{Sample size group 1 (applies to independent samples only)}
\item{n2}{Sample size group 2}
}
\description{
Helper function to convert means and standard deviations to Cohen's d.
}
\examples{


# means and standard deviations from independent samples
means.to.d(mu1 = 20, mu2 = 17.5,
           sd1 = 5, sd2 = 15,
           n2 = 30, n.ratio = 1)

# means and standard deviations from paired samples
means.to.d(mu1 = 20, mu2 = 17.5,
           sd1 = 5, sd2 = 15,
           n2 = 30, n.ratio = 1,
           paired = TRUE,
           rho.paired = 0.50)

}
