% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{means.to.etasq}
\alias{means.to.etasq}
\title{Conversion from Means and Standard Deviations to Cohen's f and Eta-squared}
\usage{
means.to.etasq(
  mu.vector,
  sd.vector,
  n.vector,
  k.covariates = 0,
  r.squared = 0,
  factor.levels = NULL,
  verbose = 1
)
}
\arguments{
\item{mu.vector}{vector of adjusted means (or estimated marginal means)
for each level of a factor.}

\item{sd.vector}{vector of unadjusted standard deviations for each level
of a factor.}

\item{n.vector}{vector of sample sizes for each level of a factor.}

\item{k.covariates}{integer; number of covariates in the ANCOVA model. The
default is \code{k.covariates = 0}, which means an
ANOVA model would be of interest.}

\item{r.squared}{explanatory power of covariates (R-squared) in the
ANCOVA model. The default is \code{r.squared = 0},
which means an ANOVA model would be of interest.}

\item{factor.levels}{integer; number of levels or groups in each factor. For
example, for two factors each having two levels or
groups use e.g. c(2, 2), for three factors each having
two levels or groups use e.g. c(2, 2, 2)}

\item{verbose}{\code{1} by default (returns results), if \code{0} no
output is printed on the console.}
}
\value{
\item{f}{Cohen's f}
\item{eta.squared}{(partial) eta-squared.}
\item{df1}{numerator degrees of freedom.}
\item{df2}{denominator degrees of freedom.}
\item{ncp}{non-centrality parameter under alternative.}
}
\description{
Calculates Cohen's f or Eta-squared for one-way ANOVA/ANCOVA. Set \code{k.cov =
0} for one-way ANOVA (without any pretest or covariate adjustment). Set
\code{k.cov > 0} in combination with \code{r.squared > 0} for one-way ANCOVA (with
pretest or covariate adjustment).
}
\details{
Note that R has a partial matching feature which allows you to specify
shortened versions of arguments, such as \code{mu} or \code{mu.vec} instead
of \code{mu.vector}, or such as \code{k} or \code{k.cov} instead of
\code{k.covariates}.
}
\examples{
means.to.etasq(mu.vector = c(0.50, 0), # marginal means
               sd.vector = c(1, 1), # unadjusted standard deviation
               n.vector = c(33, 33), # sample size (will be calculated)
               k.cov = 1, # number of covariates
               r.squared = 0.50)
}
\references{
Keppel, G., & Wickens, T. D. (2004). Design and analysis: A researcher's
handbook (4th ed.). Pearson.
}
