% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.z.test.R
\name{ncp.z.test}
\alias{ncp.z.test}
\alias{ncp.z}
\title{Finds the Non-centrality Parameter (Mean) for the Generic z-Test}
\usage{
ncp.z.test(
  power = 0.8,
  mean = NULL,
  req.sign = "+",
  sd = 1,
  null.mean = 0,
  null.sd = 1,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  plot = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{power}{statistical power \eqn{(1-\beta)}.}

\item{mean}{mean of the alternative.}

\item{req.sign}{whether 'mean' is expected to be greater '+1', less than
'-1', or within '0' the null.mean' bounds.}

\item{sd}{standard deviation of the alternative. Do not change this
value except when some sort of variance correction is
applied (e.g. as in logistic and Poisson regressions).}

\item{null.mean}{mean of the null. When alternative = "two.one.sided", the
function expects two values in the form \code{c(lower, upper)}. If a single value is provided, it is interpreted
as the absolute bound and automatically expanded to
\code{c(-value, +value)}.}

\item{null.sd}{standard deviation of the null. Do not change this value
except when some sort of correction is applied.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided", "one.sided", or "two.one.sided".
"two.one.sided" is used for equivalence and minimal
effect testing.}

\item{plot}{logical; \code{FALSE} switches off Type 1 and Type 2
error plot. \code{TRUE} by default.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{mean}{mean of the alternative distribution.}
\item{sd}{standard deviation of the alternative distribution.}
\item{null.mean}{mean of the null distribution.}
\item{null.sd}{standard deviation of the null distribution.}
\item{alpha}{type 1 error rate (user-specified).}
\item{z.alpha}{critical value(s).}
\item{beta}{type 2 error rate.}
\item{type.s}{type S error rate (only for two-tailed test).}
\item{type.m}{type M error rate (only for two-tailed test).}
\item{power}{statistical power \eqn{(1-\beta)}.}
}
\description{
Finds the non-centrality parameter (mean) for the generic z-Test with
(optional) Type 1 and Type 2 error plots.
}
\examples{
# two-sided
# power defined as the probability of observing test statistics greater than
# the positive critical value OR less than the negative critical value
ncp.z.test(mean = NULL, alpha = 0.05, alternative = "two.sided")

# one-sided
# power is defined as the probability of observing a test statistic greater
# than the critical value
ncp.z.test(mean = NULL, alpha = 0.05, alternative = "one.sided")

# equivalence
# power is defined as the probability of observing a test statistic greater
# than the upper critical value (for the lower bound) AND less than the
# lower critical value (for the upper bound)
ncp.z.test(mean = NULL, null.mean = c(-2, 2), alpha = 0.05,
             alternative = "two.one.sided")

# minimal effect testing
# power is defined as the probability of observing a test statistic greater
# than the upper critical value (for the upper bound) OR less than the lower
# critical value (for the lower bound).
ncp.z.test(mean = NULL, null.mean = c(-1, 1), alpha = 0.05,
             alternative = "two.one.sided")

}
