% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.binom.test.R
\name{power.binom.test}
\alias{power.binom.test}
\alias{power.binom}
\title{Power Analysis for the Generic Binomial Test}
\usage{
power.binom.test(
  power = NULL,
  size = NULL,
  prob = NULL,
  null.prob = 0.5,
  req.sign = "+",
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  plot = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{power}{statistical power \eqn{(1 - \beta)}; either \code{power},
\code{size}, or \code{prob} needs to be NULL (and is then
estimated).}

\item{size}{number of trials (zero or more); either \code{power}, \code{size},
or \code{prob} needs to be NULL (and is then estimated).}

\item{prob}{probability of success on each trial under alternative;
either \code{power}, \code{size} or \code{prob} needs to be NULL (and is
then estimated).}

\item{null.prob}{probability of success on each trial under null.}

\item{req.sign}{whether 'prob' is expected to be greater '+1', less than
'-1', or within '0' the null.prob' bounds.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided", "one.sided", or "two.one.sided". For
non-inferiority or superiority tests, add or subtract the
margin from the null hypothesis value and use alternative
= "one.sided".}

\item{plot}{logical; \code{FALSE} switches off Type 1 and Type 2
error plot. \code{TRUE} by default.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{power}{statistical power \eqn{(1 - \beta)}.}
\item{size}{number of trials (zero or more).}
\item{prob}{probability of success on each trial under alternative.}
\item{null.prob}{probability of success on each trial under null.}
\item{alpha}{type 1 error rate.}
\item{alternative}{direction or type of the hypothesis test.}
\item{binom.alpha}{critical value(s).}
}
\description{
Calculates power or find Probability (Non-Centrality) for the generic
binomial test with (optional) Type 1 and Type 2 error plots.
}
\examples{
# one-sided
power.binom.test(size = 200, prob = 0.6, null.prob = 0.5, alpha = 0.05,
                 alternative = "one.sided")
power.binom.test(power = 0.80, size = 200, req.sign = "+", null.prob = 0.5,
                 alpha = 0.05, alternative = "one.sided")

# two-sided
power.binom.test(size = 200, prob = 0.4, null.prob = 0.5, alpha = 0.05,
                 alternative = "two.sided")
power.binom.test(power = 0.80, size = 200, req.sign = "+", null.prob = 0.5,
                 alpha = 0.05, alternative = "two.sided")

# equivalence
power.binom.test(size = 200, prob = 0.5, null.prob = c(0.4, 0.6),
                 alpha = 0.05, alternative = "two.one.sided")
power.binom.test(power = 0.80, size = 200, req.sign = "0",
                 null.prob = c(0.4, 0.6), alpha = 0.05,
                 alternative = "two.one.sided")

}
