% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.onetwo.R
\name{power.exact.twoprops}
\alias{power.exact.twoprops}
\alias{power.exact.twoprop}
\title{Power Analysis for Testing Difference Between Two Proportions (Exact Method)}
\usage{
power.exact.twoprops(
  prob1 = NULL,
  prob2 = NULL,
  req.sign = "+",
  n.ratio = 1,
  n2 = NULL,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided"),
  paired = FALSE,
  rho.paired = 0.5,
  method = c("exact", "approximate"),
  ceil.n = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{prob1}{probability of success in the first group.}

\item{prob2}{probability of success in the second group.}

\item{req.sign}{whether estimated prob is smaller or larger than the
other (when minimum detectable prob is of interest).}

\item{n.ratio}{sample size ratio (n1 / n2).}

\item{n2}{integer; sample size for the second group.}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted as
\eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided" or "one.sided".}

\item{paired}{logical; if \code{TRUE} samples are paired. \code{FALSE}
by default.}

\item{rho.paired}{correlation between paired observations.}

\item{method}{character; whether to use "approximate" or "exact"
method. Default is \code{"exact"} (only in the
\code{power.exact.twoprops()} function).}

\item{ceil.n}{logical; \code{TRUE} rounds up sample size in each group.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the test, which is "z" or "exact".}
\item{delta}{difference between \code{prob1} and \code{prob2}}
\item{odds.ratio}{Odds-ratio \eqn{(prob1 / (1 - prob1)) /
                    (prob2 / (1 - prob2))}}
\item{size}{... (applies to paired proportions).}
\item{prob}{... (applies to paired proportions).}
\item{null.prob}{... (applies to paired proportions).}
\item{binom.alpha}{critical value(s) (applies to paired proportions).}
\item{mean}{mean of the alternative distribution.}
\item{sd}{standard deviation of the alternative distribution.}
\item{null.mean}{mean of the null distribution.}
\item{null.sd}{standard deviation of the null distribution.}
\item{alternative}{standard deviation of the null distribution.}
\item{z.alpha}{critical value(s).}
\item{power}{statistical power \eqn{(1 - \beta)}.}
\item{n}{sample size in the form of c(n1, n2) (applies to independent
proportions).}
\item{n.total}{total sample size (applies to independent proportions).}
\item{n.paired}{paired sample size (applies to paired proportions).}
}
\description{
Calculates power or sample size (only one can be NULL at a time) for two
proportions using the exact method. The function is a wrapper for
\code{power.exact.mcnemar} (if \code{paired} == TRUE), or \code{power.exact.fisher} (if
\code{paired} == FALSE)

Validated via G*Power and PASS documentation.
}
\examples{
  # power
  power.exact.twoprops(prob1 = 0.70, prob2 = 0.60,
                       alpha = 0.05, n2 = 500,
                       alternative = "one.sided")

  power.exact.twoprops(prob1 = 0.70, prob2 = 0.60,
                       alpha = 0.05, n2 = 500,
                       alternative = "one.sided",
                       paired = TRUE)

  # sample size
  power.exact.twoprops(prob1 = 0.70, prob2 = 0.60,
                       alpha = 0.05, power = 0.80,
                       alternative = "one.sided")

  power.exact.twoprops(prob1 = 0.70, prob2 = 0.60,
                       alpha = 0.05, power = 0.80,
                       alternative = "one.sided",
                       paired = TRUE)

}
\references{
Bulus, M., & Polat, C. (2023). pwrss R paketi ile istatistiksel guc
analizi [Statistical power analysis with pwrss R package]. \emph{Ahi Evran
Universitesi Kirsehir Egitim Fakultesi Dergisi, 24}(3), 2207-2328.
https://doi.org/10.29299/kefad.1209913
}
