% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.f.test.R
\name{power.f.test}
\alias{power.f.test}
\alias{power.f}
\title{Statistical Power for the Generic F-Test}
\usage{
power.f.test(
  power = NULL,
  ncp = NULL,
  null.ncp = 0,
  df1,
  df2,
  alpha = 0.05,
  plot = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{power}{statistical power \eqn{(1 - \beta)}; either \code{power} or \code{ncp}
needs to be NULL (and is then estimated).}

\item{ncp}{non-centrality parameter for the alternative; either \code{power}
or \code{ncp} needs to be NULL (and is then estimated).}

\item{null.ncp}{non-centrality parameter for the null.}

\item{df1}{integer; numerator degrees of freedom.}

\item{df2}{integer; denominator degrees of freedom.}

\item{alpha}{type 1 error rate, defined as the probability of incorrectly
rejecting a true null hypothesis, denoted as \eqn{\alpha}.}

\item{plot}{logical; \code{FALSE} switches off Type 1 and Type 2 error
plot. \code{TRUE} by default.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and definitions), if \code{0} no output is printed
on the console.}

\item{utf}{logical; whether the output should show Unicode characters
(if encoding allows for it). \code{FALSE} by default.}
}
\value{
\item{power}{statistical power \eqn{(1-\beta)}.}
\item{ncp}{non-centrality parameter under alternative.}
\item{null.ncp}{non-centrality parameter under null.}
\item{df1}{numerator degrees of freedom.}
\item{df2}{denominator degrees of freedom.}
\item{f.alpha}{critical value(s).}
}
\description{
Determines the power or the non-centrality parameter for the generic F-Test
with (optional) Type 1 and Type 2 error plots.
}
\examples{
# power is defined as the probability of observing a test statistics greater
# than the critical value
power.f.test(ncp = 1, df1 = 4, df2 = 100, alpha = 0.05)
power.f.test(power = 0.80, df1 = 4, df2 = 100, alpha = 0.05)

}
