% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.lp.test.R
\name{power.lp.test}
\alias{power.lp.test}
\alias{power.lp}
\title{Statistical Power for the Lambda-Prime Distribution}
\usage{
power.lp.test(
  power = NULL,
  ncp = NULL,
  req.sign = "+",
  null.ncp = 0,
  df = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  plot = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{power}{statistical power \eqn{(1 - \beta)}; either \code{power},
\code{ncp} or \code{df} needs to be NULL (and is then estimated).}

\item{ncp}{non-centrality parameter for the alternative; either
\code{power}, \code{ncp} or \code{df} needs to be NULL (and is then
estimated).}

\item{req.sign}{whether \code{ncp} is expected to be greater '+1', less than
'-1', or within '0' the \code{null.ncp} bounds; only relevant
if \code{ncp} is to be estimated.}

\item{null.ncp}{non-centrality parameter for the null. When alternative =
"two.one.sided", the function expects two values in the
form \code{c(lower, upper)}. If a single value is provided, it
is interpreted as the absolute bound and automatically
expanded to \code{c(-value, +value)}.}

\item{df}{degrees of freedom; either \code{power}, \code{ncp} or \code{df} needs
to be NULL (and is then estimated).}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided", "one.sided", or "two.one.sided".
"two.one.sided" is used for equivalence and minimal
effect testing.}

\item{plot}{logical; \code{FALSE} switches off Type 1 and Type 2
error plot. \code{TRUE} by default.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{power}{statistical power \eqn{(1 - \beta)}.}
\item{ncp}{non-centrality parameter under alternative.}
\item{null.ncp}{non-centrality parameter under null.}
\item{df}{degrees of freedom.}
\item{alpha}{type 1 error rate (user-specified).}
\item{alternative}{the direction or type of the hypothesis test.}
\item{t.alpha}{critical value(s).}
\item{beta}{type 2 error rate.}
\item{type.s}{type S error rate (only for two-tailed test).}
\item{type.m}{type M error rate (only for two-tailed test).}
}
\description{
Determines the power, the non-centrality parameter, or the degrees of
freedom for the lambda-prime distribution with (optional) Type 1 and Type 2
error plots.
}
\examples{
# two-sided
# power defined as the probability of observing test statistics greater
# than the positive critical value OR less than the negative critical value
power.lp.test(ncp = 1.96, df = 100, alpha = 0.05, alternative = "two.sided")
power.lp.test(power = 0.80, df = 100, alpha = 0.05, alternative = "two.sided")

# one-sided
# power is defined as the probability of observing a test statistic greater
# than the critical value
power.lp.test(ncp = 1.96, df = 100, alpha = 0.05, alternative = "one.sided")
power.lp.test(power = 0.80, df = 100, alpha = 0.05, alternative = "one.sided")

# equivalence
# power is defined as the probability of observing a test statistic greater
# than the upper critical value (for the lower bound) AND less than the
# lower critical value (for the upper bound)
power.lp.test(ncp = 0, null.ncp = c(-2, 2), df = 100, alpha = 0.05,
              alternative = "two.one.sided")
power.lp.test(power = 0.80, req.sign = "0", null.ncp = c(-2, 2),
              df = 100, alpha = 0.05, alternative = "two.one.sided")

# minimal effect testing
# power is defined as the probability of observing a test statistic greater
# than the upper critical value (for the upper bound) OR less than the lower
# critical value (for the lower bound).
power.lp.test(ncp = 2, null.ncp = c(-1, 1), df = 100, alpha = 0.05,
              alternative = "two.one.sided")
power.lp.test(power = 0.80, req.sign = "+", null.ncp = c(-1, 1),
              df = 100, alpha = 0.05, alternative = "two.one.sided")

}
