% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{power.t.contrast}
\alias{power.t.contrast}
\alias{pwrss.t.contrast}
\title{Power Analysis for One-, Two-, Three-Way ANCOVA Contrasts and Multiple
Comparisons (T-Tests)}
\usage{
power.t.contrast(
  mu.vector,
  sd.vector,
  n.vector = NULL,
  p.vector = NULL,
  contrast.vector,
  r.squared = 0,
  k.covariates = 1,
  power = NULL,
  alpha = 0.05,
  tukey.kramer = FALSE,
  ceil.n = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{mu.vector}{vector; adjusted means (or estimated marginal means)
for each level of a factor.}

\item{sd.vector}{vector; unadjusted standard deviations for each level
of a factor.}

\item{n.vector}{vector; sample sizes for each level of a factor.}

\item{p.vector}{vector; proportion of total sample size in each level
of a factor. These proportions should sum to one.}

\item{contrast.vector}{vector; a single contrast in the form of a vector
with as many elements as number of levels or groups
(or cells in factorial designs). Ignored when 'x' is
specified.}

\item{r.squared}{explanatory power of covariates (R-squared) in the
ANCOVA model.}

\item{k.covariates}{Number of covariates in the ANCOVA model.}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted
as \eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted
as \eqn{\alpha}.}

\item{tukey.kramer}{logical; \code{FALSE} by default. If \code{TRUE}
adjustments will be made to control Type 1 error.}

\item{ceil.n}{logical; \code{TRUE} by default. If \code{FALSE}
sample sizes in each cell are NOT rounded up.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE}
by default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the statistical test (T-Test).}
\item{psi}{contrast-weighted mean difference.}
\item{d}{contrast-weighted standardized mean difference.}
\item{df}{degrees of freedom.}
\item{t.alpha}{critical values.}
\item{ncp}{non-centrality parameter for the alternative.}
\item{null.ncp}{non-centrality parameter for the null.}
\item{power}{statistical power \eqn{(1-\beta)}.}
\item{n.vector}{sample sizes for each level of a factor.}
\item{n.total}{total sample size.}
}
\description{
Calculates power or sample size for a single one-, two-, three-Way ANCOVA
contrast.

Formulas are validated using examples and tables in Shieh (2017).
}
\details{
Note that R has a partial matching feature which allows you to specify
shortened versions of arguments, such as \code{mu} or \code{mu.vec} instead
of \code{mu.vector}, or such as \code{k} or \code{k.cov} instead of
\code{k.covariates}.
}
\examples{
# dummy coding example (uses the first contrast from a three-level- / two-contrasts-design)
contrast.object <- factorial.contrasts(factor.levels = 3, coding = "treatment", verbose = 0)
contrast.vector <- contrast.object[["contrast.matrix"]][1, ]
power.t.contrast(mu.vector = c(0.15, 0.30, 0.20),
                 sd.vector = c(1,    1,    1),
                 p.vector  = c(1/3,  1/3,  1/3),
                 r.squared = 0.50, k.covariates = 1,
                 contrast.vector = contrast.vector,
                 power = 0.80, alpha = 0.05)

}
\references{
Shieh, G. (2017). Power and sample size calculations for contrast analysis
in ANCOVA. \emph{Multivariate Behavioral Research, 52}(1), 1-11.
https://doi.org/10.1080/00273171.2016.1219841
}
