% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{power.t.contrasts}
\alias{power.t.contrasts}
\alias{pwrss.t.contrasts}
\title{Power Analysis for One-, Two-, Three-Way ANCOVA Contrasts and Multiple
Comparisons (T-Tests)}
\usage{
power.t.contrasts(
  x = NULL,
  mu.vector = NULL,
  sd.vector = NULL,
  n.vector = NULL,
  p.vector = NULL,
  r.squared = 0,
  k.covariates = 1,
  contrast.matrix = NULL,
  power = NULL,
  alpha = 0.05,
  adjust.alpha = c("none", "tukey", "bonferroni", "holm", "hochberg", "hommel", "BH",
    "BY", "fdr"),
  ceil.n = TRUE,
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{x}{object; an object returned from the
\code{power.f.ancova.shieh()} function.}

\item{mu.vector}{vector; adjusted means (or estimated marginal means)
for each level of a factor. Ignored when 'x' is
specified.}

\item{sd.vector}{vector; unadjusted standard deviations for each level
of a factor. Ignored when 'x' is specified.}

\item{n.vector}{vector; sample sizes for each level of a factor.
Ignored when 'x' is specified.}

\item{p.vector}{vector; proportion of total sample size in each level
of a factor. These proportions should sum to one.
Ignored when 'x' is specified.}

\item{r.squared}{explanatory power of covariates (R-squared) in the
ANCOVA model. Ignored when 'x' is specified.}

\item{k.covariates}{Number of covariates in the ANCOVA model. Ignored
when 'x' is specified.}

\item{contrast.matrix}{vector or matrix; contrasts should not be confused
with the model (design) matrix. Rows of contrast
matrix indicate independent vector of contrasts
summing to zero. The default contrast matrix is
constructed using deviation coding scheme (a.k.a.
effect coding). Columns in the contrast matrix
indicate number of levels or groups (or cells in
factorial designs). Ignored when 'x' is specified.}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted
as \eqn{1 - \beta}. Ignored when 'x' is specified.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted
as \eqn{\alpha}. Note that this should be specified
even if 'x' is specified. The 'alpha' value within
the 'x' object pertains to the omnibus test, NOT the
test of contrasts.}

\item{adjust.alpha}{character; one of the methods in c("none", "tukey",
"bonferroni", "holm", "hochberg", "hommel", "BH",
"BY", "fdr") to control Type 1 error. See
\code{?stats::p.adjust}.}

\item{ceil.n}{logical; \code{TRUE} by default. If \code{FALSE}
sample sizes in each cell are NOT rounded up.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE}
by default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the statistical test (T-Test).}
\item{contrast}{contrast number (one contrast per line).}
\item{comparison}{which factor levels are compared (one contrast per line).}
\item{psi}{contrast-weighted mean difference (one contrast per line).}
\item{d}{contrast-weighted standardized mean difference (one contrast per line).}
\item{ncp}{non-centrality parameter for the alternative (one contrast per line).}
\item{df}{degrees of freedom (one contrast per line).}
\item{t.alpha}{critical values (one contrast per line).}
\item{n.total}{total sample size (one contrast per line).}
\item{power}{statistical power \eqn{(1-\beta)} (one contrast per line).}
}
\description{
Calculates power or sample size for one-, two-, three-Way ANCOVA contrasts
and multiple comparisons. The \code{power.t.contrasts()} function permits
to test multiple contrasts (multiple comparisons) and also allows adjustment
to alpha due to multiple testing. Furthermore, \code{power.t.contrasts()}
accepts an object returned from the \code{power.f.ancova.shieh()} function
for convenience. Beware that, in this case, all other arguments are ignored
except \code{alpha} and \code{adjust.alpha}.

Formulas are validated using examples and tables in Shieh (2017).
}
\details{
Note that R has a partial matching feature which allows you to specify
shortened versions of arguments, such as \code{mu} or \code{mu.vec} instead
of \code{mu.vector}, or such as \code{k} or \code{k.cov} instead of
\code{k.covariates}.
}
\examples{
# see `?pwrss::power.f.ancova.shieh` for further examples

# dummy coding example
contrast.object <- factorial.contrasts(factor.levels = 3, coding = "treatment", verbose = 0)
contrast.matrix <- contrast.object[["contrast.matrix"]]
power.t.contrasts(mu.vector = c(0.15, 0.30, 0.20),
                  sd.vector = c(1,    1,    1),
                  p.vector  = c(1/3,  1/3,  1/3),
                  r.squared = 0.50, k.covariates = 1,
                  contrast.matrix = contrast.matrix,
                  power = 0.80,
                  alpha = 0.05, adjust.alpha = "holm")

}
\references{
Shieh, G. (2017). Power and sample size calculations for contrast analysis
in ANCOVA. \emph{Multivariate Behavioral Research, 52}(1), 1-11.
https://doi.org/10.1080/00273171.2016.1219841
}
