% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{q.to.cors}
\alias{q.to.cors}
\title{Conversion from a Cohen's q to a correlation difference}
\usage{
q.to.cors(q, rho1 = NULL, rho2 = NULL, verbose = 1)
}
\arguments{
\item{q}{Cohen's q effect size.}

\item{rho1}{first correlation (either rho1 or rho2 needs to be given)}

\item{rho2}{second correlation (either rho1 or rho2 needs to be given)}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and definitions), if \code{0} no output is printed
on the console.}
}
\value{
\item{q}{Cohen's q effect size.}
\item{delta}{correlation difference: rho1 - rho2.}
\item{rho1}{first correlation.}
\item{rho2}{second correlation.}
}
\description{
Conversion from a Cohen's q to a correlation difference
}
\examples{
q.to.cors(q = 0.10, rho1 = 0.50)
q.to.cors(q = 0.30, rho1 = 0.50)
q.to.cors(q = 0.50, rho1 = 0.50)

}
