% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsq.to.f}
\alias{rsq.to.f}
\title{Conversion from R-squared to Cohen's f}
\usage{
rsq.to.f(r.squared.full, r.squared.reduced = 0, verbose = 0)
}
\arguments{
\item{r.squared.full}{R-squared for the full model.}

\item{r.squared.reduced}{R-squared for the reduced model.}

\item{verbose}{\code{1} by default (returns results), if \code{0}
no output is printed on the console.}
}
\value{
\item{f.squared}{Cohen's f-squared.}
\item{f}{Cohen's f.}
\item{r.squared.full}{R-squared for the full model.}
\item{r.squared.reduced}{R-squared for the reduced model.}
}
\description{
Helper function to convert between Cohen's f and R-squared.
}
\examples{

  rsq.to.f(r.squared.full = 0.02) # small
  rsq.to.f(r.squared.full = 0.13) # medium
  rsq.to.f(r.squared.full = 0.26) # large

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences
(2nd ed.). Lawrence Erlbaum Associates.

Selya, A. S., Rose, J. S., Dierker, L. C., Hedeker, D., & Mermelstein,
R. J. (2012). A practical guide to calculating Cohen's f2, a measure of
local effect size, from PROC MIXED. \emph{Frontiers in Psychology, 3}, Article
111. https://doi.org/10.3389/fpsyg.2012.00111
}
