% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor.to.z}
\alias{cor.to.z}
\title{Conversion from a correlation to a z-value (Fisher's z-transformation)}
\usage{
cor.to.z(rho, verbose = 1)
}
\arguments{
\item{rho}{correlation}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and defintions), if \code{0} no output is printed
on the console.}
}
\value{
\item{z}{z-value}
\item{rho}{correlation}
}
\description{
Conversion from a correlation to a z-value (Fisher's z-transformation)
}
\examples{
cor.to.z(rho = 0.1)
cor.to.z(rho = 0.5)
cor.to.z(rho = 0.9)
cor.to.z(rho = 0.99)

}
