% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{power.exact.onecor}
\alias{power.exact.onecor}
\title{Power Analysis for One-Sample Correlation (Exact)}
\usage{
power.exact.onecor(
  rho = NULL,
  req.sign = "+",
  null.rho = 0,
  n = NULL,
  n.max = 10000,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided"),
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{rho}{correlation.}

\item{req.sign}{whether estimated rho is smaller or larger than the null.rho
(when minimum detectable rho is of interest).}

\item{null.rho}{correlation when null is true. Only 0 is allowed for now.}

\item{n}{sample size.}

\item{n.max}{max. number of observations in the sample (default: 500).}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted as
\eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided" or "one.sided".}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the statistical test (Z-Test)}
\item{rho.alpha}{critical value(s).}
\item{es}{minimum detectable correlation.}
\item{power}{statistical power\eqn{(1-\beta)}.}
\item{n}{sample size.}
}
\description{
Calculates power, sample size, or minimum detectable correlation (only one
can be NULL at a time) to test a (Pearson) correlation against a constant
using exact method described in Barabesi and Greco (2002).

Formulas are validated using G*Power.
}
\examples{
# expected correlation is 0.20 and it is different from 0
# it could be 0.20 as well as -0.20
power.exact.onecor(rho = 0.20,
                   power = 0.80,
                   alpha = 0.05,
                   alternative = "two.sided")

# expected correlation is 0.20 and it is greater than 0
power.exact.onecor(rho = 0.20,
                   power = 0.80,
                   alpha = 0.05,
                   alternative = "one.sided")


}
\references{
Barabesi, L., & Greco, L. (2002). A Note on the Exact Computation of the
Student t, Snedecor F and Sample Correlation Coefficient Distribution
Functions. Journal of the Royal Statistical Society. Series D (The
Statistician), 51(1), 105–110. https://www.jstor.org/stable/3650394
}
