% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{power.exact.onecor}
\alias{power.exact.onecor}
\title{Power Analysis for One-Sample Correlation (Exact)}
\usage{
power.exact.onecor(
  rho = NULL,
  null.rho = 0,
  n = NULL,
  n.max = 500,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided"),
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{rho}{correlation.}

\item{null.rho}{correlation when null is true.}

\item{n}{sample size.}

\item{n.max}{max sample size (applies when n = NULL).}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted as
\eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided" or "one.sided".}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the statistical test (Z-Test)}
\item{rho.alpha}{critical value(s).}
\item{mde}{minimum detectable rho.}
\item{power}{statistical power\eqn{(1-\beta)}.}
\item{n}{sample size.}
}
\description{
Calculates power, sample size, and minimum detectable correlation (only one can be NULL at a time) to test a
(Pearson) correlation against a constant using exact test propsoed by Barabesi and Greco (2002).

Formulas are validated using PASS and G*Power.
}
\examples{
# expected correlation is 0.20 and it is different from 0
power.exact.onecor(rho = 0.20,
               power = 0.80,
               alpha = 0.05,
               alternative = "two.sided")

# expected correlation is 0.20 and it is greater than 0
power.exact.onecor(rho = 0.20, 
               power = 0.80,
               alpha = 0.05,
               alternative = "one.sided")
}
\references{
Barabesi & Greco (2002): A note on the exact computation of the Student t, 
Snecdor F and sample correlation coefficient distribution function. 
\emph{The Statistician, 51}(Part 1), 105-110. 
}
