% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.onetwo.R
\name{power.exact.oneprop}
\alias{power.exact.oneprop}
\title{Power Analysis for the Test of One Proportion (Exact Method)}
\usage{
power.exact.oneprop(
  prob = NULL,
  req.sign = "+",
  null.prob = 0.5,
  n = NULL,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  verbose = 1,
  utf = FALSE
)
}
\arguments{
\item{prob}{probability of success under alternative.}

\item{req.sign}{whether \code{prob} is smaller or larger than \code{null.prob}
(when minimum detectable prob is of interest).}

\item{null.prob}{probability of success under null.}

\item{n}{integer; sample size.}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted as
\eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided", "one.sided", or "two.one.sided". For
non-inferiority or superiority tests, add margin to the
null hypothesis value and use
\code{alternative = "one.sided"}.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the statistical test ("exact").}
\item{delta}{difference between \code{prob} and \code{null.prob}}
\item{odds.ratio}{Odds-ratio \eqn{(prob / (1 - prob)) /
                    (null.prob / (1 - null.prob))}}
\item{prob}{probability of success under alternative.}
\item{null.prob}{probability of success under null.}
\item{binom.alpha}{critical value(s).}
\item{power}{statistical power \eqn{(1 - \beta)}.}
\item{n}{sample size.}
}
\description{
Calculates power, sample size or effect size (only one can be NULL at a
time) for test of a proportion against a constant using the exact method.

Formulas are validated using PASS documentation.
}
\examples{
# power'
power.exact.oneprop(prob = 0.45, null.prob = 0.50,
                    alpha = 0.05, n = 500,
                    alternative = "one.sided")

# sample size
power.exact.oneprop(prob = 0.45, null.prob = 0.50,
                    alpha = 0.05, power = 0.80,
                    alternative = "one.sided")

}
\references{
Bulus, M., & Polat, C. (2023). pwrss R paketi ile istatistiksel guc
analizi [Statistical power analysis with pwrss R package]. \emph{Ahi Evran
Universitesi Kirsehir Egitim Fakultesi Dergisi, 24}(3), 2207-2328.
https://doi.org/10.29299/kefad.1209913
}
