% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.onetwo.R
\name{power.z.twoprops}
\alias{power.z.twoprops}
\alias{power.z.twoprop}
\alias{pwrss.z.2props}
\alias{pwrss.z.2prop}
\title{Power Analysis for Testing the Difference Between Two Proportions (Normal
Approximation Method)}
\usage{
power.z.twoprops(
  prob1 = NULL,
  prob2 = NULL,
  req.sign = "+",
  margin = 0,
  n.ratio = 1,
  n2 = NULL,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  arcsine = FALSE,
  correct = FALSE,
  paired = FALSE,
  rho.paired = 0.5,
  std.error = c("pooled", "unpooled"),
  ceil.n = TRUE,
  verbose = 1,
  utf = FALSE
)

power.z.twoprop(
  prob1 = NULL,
  prob2 = NULL,
  req.sign = "+",
  margin = 0,
  n.ratio = 1,
  n2 = NULL,
  power = NULL,
  alpha = 0.05,
  alternative = c("two.sided", "one.sided", "two.one.sided"),
  arcsine = FALSE,
  correct = FALSE,
  paired = FALSE,
  rho.paired = 0.5,
  std.error = c("pooled", "unpooled"),
  ceil.n = TRUE,
  verbose = 1,
  utf = FALSE
)

pwrss.z.2props(
  p1,
  p2,
  margin = 0,
  arcsin.trans = FALSE,
  kappa = 1,
  alpha = 0.05,
  alternative = c("not equal", "greater", "less", "equivalent", "non-inferior",
    "superior"),
  n2 = NULL,
  power = NULL,
  verbose = TRUE
)

pwrss.z.2prop(
  p1,
  p2,
  margin = 0,
  arcsin.trans = FALSE,
  kappa = 1,
  alpha = 0.05,
  alternative = c("not equal", "greater", "less", "equivalent", "non-inferior",
    "superior"),
  n2 = NULL,
  power = NULL,
  verbose = TRUE
)
}
\arguments{
\item{prob1}{probability of success in the first group.}

\item{prob2}{probability of success in the second group.}

\item{req.sign}{whether estimated prob is smaller or larger than the
other (when minimum detectable prob is of interest).}

\item{margin}{ignorable \code{prob1} - \code{prob2} difference. For two
one-sided tests provide lower and upper margins in the
form of \code{c(lower, upper)}.}

\item{n.ratio}{sample size ratio (n1 / n2).}

\item{n2}{integer; sample size for the second group.}

\item{power}{statistical power, defined as the probability of
correctly rejecting a false null hypothesis, denoted as
\eqn{1 - \beta}.}

\item{alpha}{type 1 error rate, defined as the probability of
incorrectly rejecting a true null hypothesis, denoted as
\eqn{\alpha}.}

\item{alternative}{character; the direction or type of the hypothesis test:
"two.sided", "one.sided", or "two.one.sided".}

\item{arcsine}{logical; whether arcsine transformation should be
applied. Note that this only applies to independent
proportions without continuity correction.}

\item{correct}{logical; whether Yates' continuity correction should be
applied to the test statistic. Ignored for the paired
test.}

\item{paired}{logical; if \code{TRUE} samples are paired. \code{FALSE}
by default.}

\item{rho.paired}{correlation between paired observations.}

\item{std.error}{character; whether to calculate standard error using
"pooled" or "unpooled" standard deviation. Ignored for
the paired test.}

\item{ceil.n}{logical; \code{TRUE} rounds up sample size in each group.}

\item{verbose}{\code{1} by default (returns test, hypotheses, and
results), if \code{2} a more detailed output is given
(plus key parameters and definitions), if \code{0} no
output is printed on the console.}

\item{utf}{logical; whether the output should show Unicode
characters (if encoding allows for it). \code{FALSE} by
default.}
}
\value{
\item{parms}{list of parameters used in calculation.}
\item{test}{type of the test, which is "z" or "exact".}
\item{power}{statistical power \eqn{(1-\beta)}.}
\item{mean}{mean of the alternative distribution.}
\item{sd}{standard deviation of the alternative distribution.}
\item{null.mean}{mean of the null distribution.}
\item{null.sd}{standard deviation of the null distribution.}
\item{z.alpha}{critical value(s).}
\item{n}{sample size in the form of c(n1, n2) (applies to independent
proportions).}
\item{n.total}{total sample size (applies to independent proportions).}
\item{n.paired}{paired sample size (applies to paired proportions).}
}
\description{
Calculates power or sample size (only one can be NULL at a time) for two
proportions using the normal approximation method.

Validated via G*Power and PASS documentation.
}
\details{
\itemize{
\item NB: The \code{pwrss.z.2props()} function is deprecated, but it will remain
available as a wrapper for the \code{power.z.twoprops()} function during
a transition period.
}
}
\examples{
  # power
  power.z.twoprops(prob1 = 0.65, prob2 = 0.60,
                   alpha = 0.05, n2 = 500,
                   alternative = "one.sided")

  # sample size
  power.z.twoprops(prob1 = 0.65, prob2 = 0.60,
                   alpha = 0.05, power = 0.80,
                   alternative = "one.sided")

}
\references{
Bulus, M., & Polat, C. (2023). pwrss R paketi ile istatistiksel guc
analizi [Statistical power analysis with pwrss R package]. \emph{Ahi Evran
Universitesi Kirsehir Egitim Fakultesi Dergisi, 24}(3), 2207-2328.
https://doi.org/10.29299/kefad.1209913
}
