% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probs.to.h}
\alias{probs.to.h}
\title{Conversion from Probability Difference to Cohen's h}
\usage{
probs.to.h(prob1, prob2 = 0.5, verbose = 1)
}
\arguments{
\item{prob1}{Probability of success in the first group, or under the
alternative hypothesis in the one-sample case).}

\item{prob2}{Probability of success in the second group, or under the null
hypothesis in the one-sample case).}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and defintions), if \code{0} no output is printed
on the console.}

\item{h}{Cohen's h effect size.}
}
\value{
\item{h}{Cohen's h effect size.}
\item{prob1}{probability of success in the first group, or under the
alternative hypothesis in the one-sample case).}
\item{prob2}{probability of success in the second group, or under the
null hypothesis in the one-sample case).}
}
\description{
Helper function to convert probability difference to Cohen's h (and vice
versa).
}
\examples{
probs.to.h(prob1 = 0.56, prob2 = 0.50)

}
