% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{z.to.cor}
\alias{z.to.cor}
\title{Conversion from a z-value to a correlation (inverse Fisher's z-transformation)}
\usage{
z.to.cor(z, verbose = 1)
}
\arguments{
\item{z}{z-value}

\item{verbose}{\code{1} by default (returns test, hypotheses, and results),
if \code{2} a more detailed output is given (plus key
parameters and defintions), if \code{0} no output is printed
on the console.}
}
\value{
\item{rho}{correlation}
\item{z}{z-value}
}
\description{
Conversion from a z-value to a correlation (inverse Fisher's z-transformation)
}
\examples{
z.to.cor(z = 0.1003353)
z.to.cor(z = 0.5493061)
z.to.cor(z = 1.4722193)
z.to.cor(z = 2.6466524)

}
